Update Q4
Previously in the solutions we had "Manual Feature Selection" while there are also statistical feature selection methods which are actually more commonly used. For that I changed this listing to "Feature Selection(Manual or Statistical)"

source: https://machinelearningmastery.com/feature-selection-with-real-and-categorical-data/
in the above link they mention correlation as an unsupervised feature selection methods. correlation belongs to the statistical portion of methods.