Added Answers to a couple of questions (#9)
* Update README.md

a short answer for question three added

* Update Q4

Previously in the solutions we had "Manual Feature Selection" while there are also statistical feature selection methods which are actually more commonly used. For that I changed this listing to "Feature Selection(Manual or Statistical)"

source: https://machinelearningmastery.com/feature-selection-with-real-and-categorical-data/
in the above link they mention correlation as an unsupervised feature selection methods. correlation belongs to the statistical portion of methods.

* Q6 short answer added from source

Added a short paragraph summarizing the detailed explanations in the answer link

* Update README.md

* Q7 answer completed

Added a more complete answer for question 7. The previous source is kept for reference as "src1"

* Update README.md

* Update README.md

* Q15 short answer added

a short paragraph summarizing the answer link added

* Update README.md

* Q15 answer completed

* Q16 answer added

added references to two 3D reconstruction method

* Q19 solution added

* Q57 fixed spacing

* adding a separate document for NLP Questions

* Q1 and Q2 added

* Q3 added

* Q4 added

* Q5 added

* Q6 added

* Q7 added

* Q8 added

* Q9 added

* Q10 added

* Q11 added

* Update README.md

* Q12 added

* Update README.md

* Q 13 added

* Q64 added

* added contribution section

---------

Co-authored-by: Sajjad Pakdamansavoji <sajjad@UIT-M-CC221.yorku.yorku.ca>
Co-authored-by: Andrew Khalel <andrew_emel@hotmail.com>